refactor(expo): share setup as a modal route, not a custom Modal
FirstShareSetupSheet used a React Native Modal with presentationStyle
"pageSheet", which looks native but misses iOS's real interactive
dismissal gesture — it didn't behave like event/[id] or list/[slug].

Converting to an expo-router modal route (presentation: "modal" in root
Stack) gives the actual UIKit sheet with swipe-to-dismiss, proper back
button, and matches every other modal screen in the app.

- Create apps/expo/src/app/share-setup.tsx (the form, no Modal wrapper)
- Register with presentation: "modal" in app/_layout.tsx
- Simplify useShareMyList to call router.push("/share-setup") instead
  of managing isSetupSheetVisible local state; returning users still
  get native share directly via openNativeShare
- Remove all three FirstShareSetupSheet mount sites (feed/_layout,
  feed/index, settings/account) — the modal route replaces them
- Delete FirstShareSetupSheet.tsx

The share completion flow (completeSetup mutation → native share sheet
→ dismiss) is inlined in the route screen.
