+++ apps/expo/src/hooks/useShareMyList.ts
@@ -1,5 +1,6 @@
-import { useCallback, useState } from "react";
+import { useCallback } from "react";
 import { Platform, Share } from "react-native";
+import { router } from "expo-router";
 import { useUser } from "@clerk/clerk-expo";
 import { useQuery } from "convex/react";
 
@@ -9,19 +10,13 @@ import Config from "~/utils/config";
 import { logError } from "~/utils/errorLogging";
 
 /**
- * Wraps Share.share for the user's personal Soon List. On the first share
- * attempt (when hasSharedListBefore is false/undefined), opens the
- * FirstShareSetupSheet instead of calling Share.share directly. After the
- * sheet commits or is dismissed via "Share now", the actual share sheet
- * is opened.
- *
- * Swipe-to-dismiss on the sheet does NOT trigger a share and does NOT
- * set hasSharedListBefore — the sheet will reappear on next attempt.
+ * Kicks off a share of the user's personal Soonlist. On the first share
+ * attempt (hasSharedListBefore false/undefined), routes to the share-setup
+ * modal screen. Returning users skip setup and get the native share sheet.
  */
 export function useShareMyList() {
   const { user } = useUser();
   const currentUser = useQuery(api.users.getCurrentUser);
-  const [isSetupSheetVisible, setSetupSheetVisible] = useState(false);
 
   const openNativeShare = useCallback(async () => {
     const username = user?.username ?? currentUser?.username ?? "";
@@ -39,22 +34,8 @@ export function useShareMyList() {
       void openNativeShare();
       return;
     }
-    setSetupSheetVisible(true);
+    router.push("/share-setup");
   }, [currentUser?.hasSharedListBefore, openNativeShare]);
 
-  const closeSetupSheet = useCallback(() => {
-    setSetupSheetVisible(false);
-  }, []);
-
-  const closeSetupSheetAndShare = useCallback(async () => {
-    setSetupSheetVisible(false);
-    await openNativeShare();
-  }, [openNativeShare]);
-
-  return {
-    requestShare,
-    isSetupSheetVisible,
-    closeSetupSheet,
-    closeSetupSheetAndShare,
-  };
+  return { requestShare };
 }
